Set disabled & window.confirm

+++ app/components/modal/MatterCardDetail.tsx
@@ -131,6 +131,17 @@ export const MatterCardDetailModal = ({
   const handleUpdateMatterInfo = async (
     updatedMatterInfo: UpdatedMatterInfoType
   ) => {
+    const checkUpdate = window.confirm(
+      `案件[${updatedMatterInfo.title}]を更新しますか？`
+    );
+    if (!checkUpdate) {
+      closeModal();
+      return;
+    }
+    matterInfo.is_fixed = window.confirm(
+      `案件[${updatedMatterInfo.title}]を確定にしますか？\n確定後は経理の確認に入るため、変更できません。`
+    );
+
     matterInfo.title = updatedMatterInfo.title;
     matterInfo.category = updatedMatterInfo.category;
     matterInfo.team = updatedMatterInfo.team;
@@ -174,14 +185,24 @@ export const MatterCardDetailModal = ({
       }
     }
 
+    alert(`案件[${matterInfo.title}]を更新しました。`);
     closeModal();
   };
 
   const handleDeleteMatterInfo = async () => {
+    const checkDelete = window.confirm(
+      `案件[${matterInfo.title}]を削除してよろしいですか？`
+    );
+    if (!checkDelete) {
+      alert(`案件[${matterInfo.title}]の削除を中止しました。`);
+      closeModal();
+      return;
+    }
     for (const costInfo of costInfoInCardList) {
       await deleteCostInfoInSupabase(costInfo.id);
     }
     await deleteMatterInfoInSupabase(matterInfo.id);
+    alert(`案件[${matterInfo.title}]を削除しました。`);
     closeModal();
   };
 
@@ -228,12 +249,15 @@ export const MatterCardDetailModal = ({
           withAsterisk
           label="案件名"
           placeholder="案件名をご記入ください。"
+          required
+          disabled={matterInfo.is_fixed!}
           {...form.getInputProps("title")}
         />
         <TextInput
           withAsterisk
           label="取引先"
           placeholder="取引先をご記入ください。"
+          disabled={matterInfo.is_fixed!}
           {...form.getInputProps("billing_address")}
         />
         <Select
@@ -241,6 +265,7 @@ export const MatterCardDetailModal = ({
           placeholder="案件の分類をご記入ください。"
           data={categoryList}
           required
+          disabled={matterInfo.is_fixed!}
           {...form.getInputProps("category")}
         />
         <Select
@@ -248,6 +273,7 @@ export const MatterCardDetailModal = ({
           placeholder="案件を担当するチームを選択ください。"
           data={teamList}
           required
+          disabled={matterInfo.is_fixed!}
           {...form.getInputProps("team")}
         />
         <NumberInput
@@ -258,12 +284,14 @@ export const MatterCardDetailModal = ({
           allowNegative={false}
           allowDecimal={false}
           thousandSeparator=","
+          disabled={matterInfo.is_fixed!}
           {...form.getInputProps("amount")}
         />
         <DateInput
           label="案件開始日"
           required
           placeholder="案件を開始した日付をご入力ください。"
+          disabled={matterInfo.is_fixed!}
           valueFormat="YYYY/MM/DD"
           value={startDate}
           onChange={(value) =>
@@ -273,6 +301,7 @@ export const MatterCardDetailModal = ({
         <DateInput
           label="請求日"
           placeholder="案件に対する報酬を請求する日付をご入力ください。"
+          disabled={matterInfo.is_fixed!}
           valueFormat="YYYY/MM/DD"
           value={invoiceDate}
           onChange={(value) =>
@@ -282,6 +311,7 @@ export const MatterCardDetailModal = ({
         <DateInput
           label="振込期限"
           placeholder="案件の報酬の振込期限をご入力ください。"
+          disabled={matterInfo.is_fixed!}
           valueFormat="YYYY/MM/DD"
           value={periodDate}
           onChange={(value) =>
@@ -290,6 +320,7 @@ export const MatterCardDetailModal = ({
         />
         <TextInput
           label="説明"
+          disabled={matterInfo.is_fixed!}
           placeholder="案件に追加の説明があればご記入ください。"
           {...form.getInputProps("description")}
         />
@@ -311,21 +342,24 @@ export const MatterCardDetailModal = ({
               padding="lg"
               aria-label="コスト"
             >
-              <div className="flex justify-between w-full mb-2">
-                <div>コスト{index + 1}</div>
-                <div
-                  className="h-full px-4 text-lg hover:cursor-pointer w-4 ml-auto items-center justify-center"
-                  onClick={() => handleRemoveCost(cost.id)}
-                >
-                  <FaRegTrashAlt />
+              {!matterInfo.is_fixed ? (
+                <div className="flex justify-between w-full mb-2">
+                  <div>コスト{index + 1}</div>
+                  <div
+                    className="h-full px-4 text-lg hover:cursor-pointer w-4 ml-auto items-center justify-center"
+                    onClick={() => handleRemoveCost(cost.id)}
+                  >
+                    <FaRegTrashAlt />
+                  </div>
                 </div>
-              </div>
+              ) : null}
               <div className="flex-grow">
                 <div className="flex pb-2">
                   <Group gap="sm" className="flex-grow w-full">
                     <TextInput
                       placeholder="品名をご記入ください。"
                       className="flex-grow"
+                      disabled={matterInfo.is_fixed!}
                       value={cost.name}
                       onChange={(event) =>
                         setCostInfoInCardList(
@@ -342,6 +376,7 @@ export const MatterCardDetailModal = ({
                       placeholder="品目を選択ください。"
                       data={itemList}
                       required
+                      disabled={matterInfo.is_fixed!}
                       value={cost.item}
                       onChange={(value) =>
                         setCostInfoInCardList(
@@ -356,6 +391,7 @@ export const MatterCardDetailModal = ({
                     <NumberInput
                       placeholder="¥0"
                       className="flex-grow"
+                      disabled={matterInfo.is_fixed!}
                       value={cost.price}
                       prefix="¥"
                       allowNegative={false}
@@ -378,6 +414,7 @@ export const MatterCardDetailModal = ({
                     <DateInput
                       className="flex-grow"
                       placeholder="支払い期限をご記入ください。"
+                      disabled={matterInfo.is_fixed!}
                       valueFormat="YYYY/MM/DD"
                       onChange={(event) => {
                         const dateString = event
@@ -395,6 +432,7 @@ export const MatterCardDetailModal = ({
                     <TextInput
                       placeholder="支払い先の名前をご記入ください。"
                       className="flex-grow"
+                      disabled={matterInfo.is_fixed!}
                       value={cost.payment_target}
                       onChange={(event) =>
                         setCostInfoInCardList(
@@ -414,6 +452,7 @@ export const MatterCardDetailModal = ({
                       placeholder="支払いの通知方法を選択ください。"
                       data={certificateList}
                       required
+                      disabled={matterInfo.is_fixed!}
                       value={cost.certificate}
                       onChange={(value) =>
                         setCostInfoInCardList(
@@ -427,6 +466,7 @@ export const MatterCardDetailModal = ({
                     />
                     <Checkbox
                       label="源泉徴収あり"
+                      disabled={matterInfo.is_fixed!}
                       key={form.key("withholding")}
                       {...form.getInputProps("withholding", {
                         type: "checkbox",
@@ -450,6 +490,7 @@ export const MatterCardDetailModal = ({
                   <TextInput
                     placeholder="コメントがあればご記入ください。"
                     className="flex-grow w-full"
+                    disabled={matterInfo.is_fixed!}
                     value={cost.comment ? cost.comment : ""}
                     onChange={(event) =>
                       setCostInfoInCardList(
@@ -466,34 +507,38 @@ export const MatterCardDetailModal = ({
             </Card>
           )
         )}
-        <Button
-          type="button"
-          fullWidth
-          className="mt-4"
-          color="dark"
-          variant="outline"
-          rightSection={<CiSquarePlus />}
-          onClick={handleAddCost}
-        >
-          コスト追加
-        </Button>
+        {!matterInfo.is_fixed ? (
+          <Button
+            type="button"
+            fullWidth
+            className="mt-4"
+            color="dark"
+            variant="outline"
+            rightSection={<CiSquarePlus />}
+            onClick={handleAddCost}
+          >
+            コスト追加
+          </Button>
+        ) : null}
 
-        <div className="border-spacing-3 h-6"></div>
-        <div className="flex justify-between">
-          <Group justify="flex-end" mt="md">
-            <Button type="button" color="gray" onClick={handleDeleteMatterInfo}>
-              削除
-            </Button>
-          </Group>
-          <Group justify="flex-end" mt="md">
-            <Button type="button" color="pink" onClick={closeModal}>
-              キャンセル
-            </Button>
-            <Button type="submit" color="green">
-              更新
-            </Button>
-          </Group>
-        </div>
+        {!matterInfo.is_fixed ? (
+          <div className="flex justify-between mt-6">
+            <Group justify="flex-end" mt="md">
+              <Button
+                type="button"
+                color="gray"
+                onClick={handleDeleteMatterInfo}
+              >
+                削除
+              </Button>
+            </Group>
+            <Group justify="flex-end" mt="md">
+              <Button type="submit" color="green">
+                更新
+              </Button>
+            </Group>
+          </div>
+        ) : null}
       </form>
     </Modal>
   );
